--- src/main/java/adoop/Timer.java
@@ -1,5 +1,0 @@
-package adoop;
-
-public class Timer {
-
-}
