+++ src/main/java/ao/adoop/io/MutipleInputs.java
@@ -1,9 +1,9 @@
-package io;
+package ao.adoop.io;
 
 import java.nio.file.Path;
 
-import adoop.Job;
-import adoop.Mapper;
+import ao.adoop.mapreduce.Job;
+import ao.adoop.mapreduce.Mapper;
 
 public class MutipleInputs {
