+++ src/main/java/ao/adoop/mapreduce/CommandLineInterface.java
@@ -1,4 +1,4 @@
-package adoop;
+package ao.adoop.mapreduce;
 
 public class CommandLineInterface implements UserInterface {
