+++ src/main/java/ao/adoop/mapreduce/Configuration.java
@@ -1,4 +1,4 @@
-package adoop;
+package ao.adoop.mapreduce;
 
 public class Configuration {
