+++ src/main/java/ao/adoop/mapreduce/Context.java
@@ -1,4 +1,4 @@
-package adoop;
+package ao.adoop.mapreduce;
 
 import java.util.ArrayList;
 import java.util.HashMap;
