+++ src/main/java/ao/adoop/mapreduce/InvalidReducerException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package ao.adoop.mapreduce;
 
 public class InvalidReducerException extends Exception{
 	private static final long serialVersionUID = 4862618313133993820L;
