+++ src/main/java/ao/adoop/mapreduce/Timer.java
@@ -1,0 +1,5 @@
+package ao.adoop.mapreduce;
+
+public class Timer {
+
+}
