+++ src/main/java/ao/adoop/ui/CommandHandler.java
@@ -1,4 +1,4 @@
-package commandparser;
+package ao.adoop.ui;
 
 import java.util.ArrayList;
 import java.util.Arrays;
