+++ src/test/java/adooptest/DataLoaderTest.java
@@ -9,7 +9,7 @@ import java.util.ArrayList;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import io.DataLoader;
+import ao.adoop.io.DataLoader;
 
 public class DataLoaderTest {
 	DataLoader loader = new DataLoader();
